requirements.txt
altair==4.2.2
hvac==2.3.0
json_fix==1.0.0
pandas==2.2.2
prettytable==3.11.0
psutil==6.0.0
pydantic==2.8.2
pytest==8.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
streamlit==1.37.1
streamlit_autorefresh==1.0.1
streamlit_pills==0.3.0
typing_extensions==4.12.2
urllib3==2.2.2